@@ -1,10 +1,10 @@
 source 'https://rubygems.org'
-ruby '2.1.10'
+ruby '2.7.3'
 
-gem 'rails', '4.1.4'
-gem 'sass-rails', '~> 4.0.3'
-gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.0.0'
+gem 'rails'
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
 
 gem 'rack-cors'
 
@@ -14,7 +14,7 @@ gem 'jquery-ui-rails'
 gem "active_model_serializers"
 
 gem 'devise'
-gem 'activeadmin', github: 'gregbell/active_admin'
+gem 'activeadmin'
 
 gem 'paperclip'
 gem 'aws-sdk'
