@@ -1,4 +1,4 @@
-class AddImageToEntry < ActiveRecord::Migration
+class AddImageToEntry < ActiveRecord::Migration[4.2]
   def change
   	change_table :entries do |t|
       t.attachment :image
