@@ -1,4 +1,4 @@
-class GiveCategoriesToEntries < ActiveRecord::Migration
+class GiveCategoriesToEntries < ActiveRecord::Migration[4.2]
   def change
   	create_table :categories_entries do |t|
   		t.belongs_to :category
