@@ -1,4 +1,4 @@
-class AddAudioToEntries < ActiveRecord::Migration
+class AddAudioToEntries < ActiveRecord::Migration[4.2]
   def change
   	change_table :entries do |t|
       t.attachment :audio
