@@ -1,4 +1,4 @@
-class CreateImageCredits < ActiveRecord::Migration
+class CreateImageCredits < ActiveRecord::Migration[4.2]
   def change
     create_table :image_credits do |t|
       t.string :attribution_text
