@@ -1,4 +1,4 @@
-class LinkImageCreditToEntry < ActiveRecord::Migration
+class LinkImageCreditToEntry < ActiveRecord::Migration[4.2]
   def change
     add_column :image_credits, :entry_id, :integer
   end
