@@ -1,4 +1,4 @@
-class DropTranslationFromEntry < ActiveRecord::Migration
+class DropTranslationFromEntry < ActiveRecord::Migration[4.2]
   def change
     remove_column :entries, :translation
   end
